Extract Scaler interface
