+++ java/compaction/compaction-task-creation/src/main/java/sleeper/compaction/task/creation/RunTasks.java
@@ -49,6 +49,7 @@ import static sleeper.configuration.properties.instance.CdkDefinedInstanceProper
 import static sleeper.configuration.properties.instance.CdkDefinedInstanceProperty.COMPACTION_JOB_QUEUE_URL;
 import static sleeper.configuration.properties.instance.CdkDefinedInstanceProperty.COMPACTION_TASK_EC2_DEFINITION_FAMILY;
 import static sleeper.configuration.properties.instance.CdkDefinedInstanceProperty.COMPACTION_TASK_FARGATE_DEFINITION_FAMILY;
+import static sleeper.configuration.properties.instance.CdkDefinedInstanceProperty.CONFIG_BUCKET;
 import static sleeper.configuration.properties.instance.CommonProperty.FARGATE_VERSION;
 import static sleeper.configuration.properties.instance.CommonProperty.SUBNETS;
 import static sleeper.configuration.properties.instance.CompactionProperty.COMPACTION_ECS_LAUNCHTYPE;
@@ -65,6 +66,7 @@ public class RunTasks {
     private static final Logger LOGGER = LoggerFactory.getLogger(RunTasks.class);
 
     private final AmazonECS ecsClient;
+    private final InstanceProperties instanceProperties;
     private final String s3Bucket;
     private final String sqsJobQueueUrl;
     private final String clusterName;
@@ -83,42 +85,42 @@ public class RunTasks {
             AmazonECS ecsClient,
             AmazonS3 s3Client,
             AmazonAutoScaling asClient,
-            String s3Bucket) {
-        this(sqsClient, ecsClient, s3Client, asClient, s3Bucket, QueueMessageCount.withSqsClient(sqsClient),
-                (clusterName) -> CommonJobUtils.getNumPendingAndRunningTasks(clusterName, ecsClient));
+            InstanceProperties instanceProperties) {
+        this(sqsClient, ecsClient, s3Client, asClient, instanceProperties, QueueMessageCount.withSqsClient(sqsClient),
+                (clusterName) -> CommonJobUtils.getNumPendingAndRunningTasks(clusterName, ecsClient),
+                createEC2Scaler(instanceProperties, asClient, ecsClient));
     }
 
     public RunTasks(AmazonSQS sqsClient,
             AmazonECS ecsClient,
             AmazonS3 s3Client,
             AmazonAutoScaling asClient,
-            String s3Bucket,
+            InstanceProperties instanceProperties,
             QueueMessageCount.Client queueMessageCount,
-            TaskCounts taskCounts) {
+            TaskCounts taskCounts,
+            Scaler scaler) {
+        this.instanceProperties = instanceProperties;
         this.ecsClient = ecsClient;
-        this.s3Bucket = s3Bucket;
         this.queueMessageCount = queueMessageCount;
         this.taskCounts = taskCounts;
-
-        InstanceProperties instanceProperties = new InstanceProperties();
-        instanceProperties.loadFromS3(s3Client, s3Bucket);
-        String autoScalingGroupName;
+        this.scaler = scaler;
+        this.s3Bucket = instanceProperties.get(CONFIG_BUCKET);
         this.sqsJobQueueUrl = instanceProperties.get(COMPACTION_JOB_QUEUE_URL);
         this.clusterName = instanceProperties.get(COMPACTION_CLUSTER);
         this.containerName = COMPACTION_CONTAINER_NAME;
         this.fargateTaskDefinition = instanceProperties.get(COMPACTION_TASK_FARGATE_DEFINITION_FAMILY);
         this.ec2TaskDefinition = instanceProperties.get(COMPACTION_TASK_EC2_DEFINITION_FAMILY);
-        autoScalingGroupName = instanceProperties.get(COMPACTION_AUTO_SCALING_GROUP);
-
         this.maximumRunningTasks = instanceProperties.getInt(MAXIMUM_CONCURRENT_COMPACTION_TASKS);
         this.subnets = instanceProperties.getList(SUBNETS);
         this.fargateVersion = instanceProperties.get(FARGATE_VERSION);
         this.launchType = instanceProperties.get(COMPACTION_ECS_LAUNCHTYPE);
 
-        String architecture = instanceProperties.get(COMPACTION_TASK_CPU_ARCHITECTURE).toUpperCase(Locale.ROOT);
-        Pair<Integer, Integer> requirements = Requirements.getArchRequirements(architecture, launchType,
-                instanceProperties);
+    }
 
+    private static Scaler createEC2Scaler(InstanceProperties instanceProperties, AmazonAutoScaling asClient, AmazonECS ecsClient) {
+        String launchType = instanceProperties.get(COMPACTION_ECS_LAUNCHTYPE);
+        String architecture = instanceProperties.get(COMPACTION_TASK_CPU_ARCHITECTURE).toUpperCase(Locale.ROOT);
+        Pair<Integer, Integer> requirements = Requirements.getArchRequirements(architecture, launchType, instanceProperties);
         // Bit hacky: EC2s don't give 100% of their memory for container use (OS
         // headroom, system tasks, etc.) so we have to make sure to reduce
         // the EC2 memory requirement by 5%. If we don't we end up asking for
@@ -127,13 +129,12 @@ public class RunTasks {
             requirements = Pair.of(requirements.getLeft(), (int) (requirements.getRight() * 0.95));
         }
 
-        this.scaler = new Scaler(asClient, ecsClient, autoScalingGroupName, this.clusterName,
-                requirements.getLeft(),
-                requirements.getRight());
+        return new EC2Scaler(asClient, ecsClient, instanceProperties.get(COMPACTION_AUTO_SCALING_GROUP),
+                instanceProperties.get(COMPACTION_CLUSTER), requirements.getLeft(), requirements.getRight());
     }
 
     private interface TaskCounts {
-        int getPendingAndRunning(String clusterName);
+        int getRunningAndPending(String clusterName);
     }
 
     public void run() {
@@ -157,7 +158,7 @@ public class RunTasks {
             return;
         }
         // Find out number of pending and running tasks
-        int numRunningAndPendingTasks = taskCounts.getPendingAndRunning(clusterName);
+        int numRunningAndPendingTasks = taskCounts.getRunningAndPending(clusterName);
         LOGGER.info("Number of running and pending tasks is {}", numRunningAndPendingTasks);
 
         if (numRunningAndPendingTasks >= maximumRunningTasks) {
@@ -171,7 +172,7 @@ public class RunTasks {
         if (launchType.equalsIgnoreCase("EC2")) {
             int totalTasks = numberOfTasksToCreate + numRunningAndPendingTasks;
             LOGGER.info("Total number of tasks if all launches succeed {}", totalTasks);
-            scaler.scaleTo(totalTasks);
+            scaler.scaleTo(instanceProperties.get(COMPACTION_AUTO_SCALING_GROUP), totalTasks);
         }
         TaskOverride override = createOverride(List.of(s3Bucket), containerName);
         NetworkConfiguration networkConfiguration = networkConfig(subnets);
@@ -284,7 +285,9 @@ public class RunTasks {
         AmazonS3 s3Client = AmazonS3ClientBuilder.defaultClient();
         AmazonAutoScaling asClient = AmazonAutoScalingClientBuilder.defaultClient();
         try {
-            new RunTasks(sqsClient, ecsClient, s3Client, asClient, s3Bucket)
+            InstanceProperties instanceProperties = new InstanceProperties();
+            instanceProperties.loadFromS3(s3Client, s3Bucket);
+            new RunTasks(sqsClient, ecsClient, s3Client, asClient, instanceProperties)
                     .run(numberOfTasks);
         } finally {
             sqsClient.shutdown();
